chore: rename file in the other-externals dir to match the component
