+++ src/domains/other-externals/components/generate-section.tsx
